chore(vote app): create a vote app
this is to help separate the vote model from from question model

[Chores #164749125]
