+++ question/admin.py
@@ -3,5 +3,3 @@ from django.contrib import admin
 from .models import Question
 
 admin.site.register(Question)
-
-# Register your models here.
