+++ question/models.py
@@ -17,17 +17,6 @@ class Question(models.Model):
     def __str__(self):
         return (self.title, self.body)  # pragma: no cover
 
-
-class Vote(models.Model):
-    question_id = models.ForeignKey(Question, on_delete=models.CASCADE)
-    voter_id = models.ForeignKey(User, on_delete=models.CASCADE)
-    vote = models.IntegerField()
-    date_voted = models.DateTimeField(auto_now_add=True)
-
-    def __str__(self):
-        return (self.voter_id, self.vote)  # pragma: no cover
-
-
 class Comment(models.Model):
     """This class represents the comment model."""
