+++ question/tests/test_question_crud.py
@@ -315,93 +315,4 @@ class TestQuestionViews(TestCase):
         self.assertEqual(self.qn_details.id, self.qn_db.id)
         self.assertTrue(self.qn_details.delete_status)
         self.assertEqual(self.qn_details.title, self.qn_db.title)
-
-    def test_other_users_can_upvote_on_a_question(self):
-        self.client.force_authenticate(user=self.user2)
-
-        url = f"/meetups/{int(self.qn_db.meetup_id.id)}/questions/{int(self.qn_db.id)}/upvote/"
-        response = self.client.get(
-            url, content_type="application/json")
-        self.assertEqual(response.status_code, 200)
-
-    def test_user_invalid_question_upvote(self):
-        self.client.force_authenticate(user=self.user1)
-
-        url = f"/meetups/{int(self.qn_db.meetup_id.id)}/questions/1233/upvote/"
-        response = self.client.get(
-            url, content_type="application/json")
-        self.assertEqual(response.status_code, 400)
-
-    def test_user_cannot_upvote_on_their_question(self):
-        self.client.force_authenticate(user=self.user1)
-
-        url = f"/meetups/{int(self.qn_db.meetup_id.id)}/questions/{int(self.qn_db.id)}/upvote/"
-        response = self.client.get(
-            url, content_type="application/json")
-        self.assertEqual(response.status_code, 400)
-
-    def test_user_can_cancel_upvote(self):
-        self.client.force_authenticate(user=self.user2)
-
-        url = f"/meetups/{int(self.qn_db.meetup_id.id)}/questions/{int(self.qn_db.id)}/upvote/"
-        self.client.get(
-            url, content_type="application/json")
-        response = self.client.get(
-            url, content_type="application/json")
-        self.assertEqual(response.status_code, 200)
-
-    def test_user_can_turn_downvote_to_upvote(self):
-        self.client.force_authenticate(user=self.user2)
-
-        url_1 = f"/meetups/{int(self.qn_db.meetup_id.id)}/questions/{int(self.qn_db.id)}/downvote/"
-        url_2 = f"/meetups/{int(self.qn_db.meetup_id.id)}/questions/{int(self.qn_db.id)}/upvote/"
-        self.client.get(
-            url_1, content_type="application/json")
-        response = self.client.get(
-            url_2, content_type="application/json")
-        self.assertEqual(response.status_code, 200)
-
-    def test_user_cannot_downvote_on_their_question(self):
-        self.client.force_authenticate(user=self.user1)
-
-        url = f"/meetups/{int(self.qn_db.meetup_id.id)}/questions/{int(self.qn_db.id)}/downvote/"
-        response = self.client.get(
-            url, content_type="application/json")
-        self.assertEqual(response.status_code, 400)
-
-    def test_other_users_can_downvote_on_a_question(self):
-        self.client.force_authenticate(user=self.user2)
-
-        url = f"/meetups/{int(self.qn_db.meetup_id.id)}/questions/{int(self.qn_db.id)}/downvote/"
-        response = self.client.get(
-            url, content_type="application/json")
-        self.assertEqual(response.status_code, 200)
-
-    def test_user_invalid_question_downvote(self):
-        self.client.force_authenticate(user=self.user1)
-
-        url = f"/meetups/{int(self.qn_db.meetup_id.id)}/questions/1233/downvote/"
-        response = self.client.get(
-            url, content_type="application/json")
-        self.assertEqual(response.status_code, 400)
-
-    def test_user_can_cancel_downvote(self):
-        self.client.force_authenticate(user=self.user2)
-
-        url = f"/meetups/{int(self.qn_db.meetup_id.id)}/questions/{int(self.qn_db.id)}/downvote/"
-        self.client.get(
-            url, content_type="application/json")
-        response = self.client.get(
-            url, content_type="application/json")
-        self.assertEqual(response.status_code, 200)
-
-    def test_user_can_turn_upvote_to_downvote(self):
-        self.client.force_authenticate(user=self.user2)
-
-        url_1 = f"/meetups/{int(self.qn_db.meetup_id.id)}/questions/{int(self.qn_db.id)}/upvote/"
-        url_2 = f"/meetups/{int(self.qn_db.meetup_id.id)}/questions/{int(self.qn_db.id)}/downvote/"
-        self.client.get(
-            url_1, content_type="application/json")
-        response = self.client.get(
-            url_2, content_type="application/json")
-        self.assertEqual(response.status_code, 200)
+
