+++ question/urls.py
@@ -1,7 +1,6 @@
 from django.urls import path
 
-from question.views import (Questions, OneQuestion, UpVote, DownVote,
-                            CommentList, CommentDetail)
+from question.views import (Questions, OneQuestion, CommentList, CommentDetail)
 
 urlpatterns = [
     path("<int:meetup_id>/questions/", Questions.as_view(), name="questions"),
@@ -9,16 +8,6 @@ urlpatterns = [
         "<int:meetup_id>/questions/<int:question_id>/",
         OneQuestion.as_view(),
         name="question",
-    ),
-    path(
-        "<int:meetup_id>/questions/<int:question_id>/upvote/",
-        UpVote.as_view(),
-        name="upvote",
-    ),
-    path(
-        "<int:meetup_id>/questions/<int:question_id>/downvote/",
-        DownVote.as_view(),
-        name="downvote",
     ),
     path("<int:meetup_id>/questions/<question_id>/comments",
          CommentList.as_view(), name='comment'),
