+++ vote/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+
+from .models import Vote
+
+admin.site.register(Vote)
