+++ vote/serializers.py
@@ -1,0 +1,9 @@
+from rest_framework import serializers
+
+from vote.models import Vote
+
+class VoteSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Vote
+        fields = "__all__"
+
